+++ src/net/i2p/crypto/eddsa/EdDSAPublicKey.java
@@ -117,24 +117,24 @@ public class EdDSAPublicKey implements EdDSAKey, PublicKey {
             int totlen = 12 + Abyte.length;
             byte[] rv = new byte[totlen];
             // sequence
-            rv[0] = 0x30;
-            rv[1] = (byte) ((byte) (totlen - 2) & 0xff);
+            rv[0] = (byte) 0x30;
+            rv[1] = (byte) ((int) (byte) (totlen - 2) & 0xff);
             // Algorithm Identifier
             // sequence
-            rv[2] = 0x30;
-            rv[3] = 5;
+            rv[2] = (byte) 0x30;
+            rv[3] = (byte) 5;
             // OID
             // https://msdn.microsoft.com/en-us/library/windows/desktop/bb540809%28v=vs.85%29.aspx
-            rv[4] = 0x06;
-            rv[5] = 3;
-            rv[6] = (1 * 40) + 3;
-            rv[7] = 101;
+            rv[4] = (byte) 0x06;
+            rv[5] = (byte) 3;
+            rv[6] = (byte) ((1 * 40) + 3);
+            rv[7] = (byte) 101;
             rv[8] = (byte) OID_ED25519;
             // params - absent
             // the key
-            rv[9] = 0x03; // bit string
-            rv[10] = (byte) ((byte) (1 + Abyte.length) & 0xff);
-            rv[11] = 0; // number of trailing unused bits
+            rv[9] = (byte) 0x03; // bit string
+            rv[10] = (byte) ((int) (byte) (1 + Abyte.length) & 0xff);
+            rv[11] = (byte) 0; // number of trailing unused bits
             System.arraycopy(Abyte, 0, rv, 12, Abyte.length);
             return rv;
         }
